refactor(thinking): move effort ladder package under internal/protocol
The ladder is protocol-domain knowledge, so home it at
internal/protocol/thinking. Kept as its own leaf subpackage rather than
merged into the protocol base package: the base package imports
guardrails/core and others, and folding the ladder in there would tie
typ's re-exports to that dependency surface and reintroduce the cycle
risk the leaf package exists to avoid. Import paths and doc references
updated; no behavior change.
