+++ internal/protocol/thinking/thinking.go
